Implemented review suggestions

+++ src/main/java/com/chatapp/vartalaab/service/MessageService.java
@@ -42,12 +42,15 @@ public class MessageService {
         if(offlineMessageIds.isPresent()) {
             List<String> messageIds = offlineMessageIds.get().getMessageIds();
             for (String messageId : messageIds) {
-                MessageEntity message = messageRepository.findById(messageId).orElseThrow(() -> new NoSuchElementException());
-                MessageDto messageDto = new MessageDto(message.getSender(), message.getMessage(), message.getTimestamp());
-                messages.add(messageDto.toString()); //returning JSON message in text format
-                this.forwardTheMessageToReceiver(messageDto.getSender(), new TextMessage(GeneralUtility.AckForReceived));
-                messageRepository.deleteById(messageId);
+                Optional<MessageEntity> message = messageRepository.findById(messageId);
+                if(message.isPresent()) {
+                    MessageDto messageDto = new MessageDto(message.get().getSender(), message.get().getMessage(), message.get().getTimestamp());
+                    messages.add(messageDto.toString()); //returning JSON message in text format
+                    this.sendMessageToRecipient(messageDto.getSender(), new TextMessage(GeneralUtility.ACK_FOR_RECEIVED));
+                    messageRepository.deleteById(messageId);
+                }
             }
+            offlineMessageIdsRepository.deleteById(username);
         }
         return messages;
     }
@@ -64,7 +67,7 @@ public class MessageService {
         offlineMessageIdsRepository.save(tempOfileMessageIds);
     }
 
-    public void forwardTheMessageToReceiver(String messageReceiver, TextMessage message) throws IOException {
+    public void sendMessageToRecipient(String messageReceiver, TextMessage message) throws IOException {
         for(String webSocketSessionId: userSessionService.getUserWebSocketSessionIds(messageReceiver)){
             sessionMap.get(webSocketSessionId).sendMessage(message);
             //Todo: save msg to MongoDB
